refactor(cosmo_val): extract PureEBMixin (pure_eb)
